@@ -1,7 +1,7 @@
 import { ThemeProvider, CssBaseline, createTheme } from '@mui/material'
 import { StylesProvider } from '@mui/styles'
 import theme from './app/theme'
-import { BrowserRouter } from 'react-router-dom'
+import { RouterProvider } from 'react-router-dom'
 import GlobalStyle from './app/globalStyles'
 import '@fontsource/poppins'
 import Router from './app/router/router'
@@ -10,14 +10,13 @@ import NavBar from './app/NavBar'
 const App = () => {
   return (
     <StylesProvider>
-      <BrowserRouter>
-        <GlobalStyle />
-        <ThemeProvider theme={createTheme(theme)}>
-          <CssBaseline />
+      <GlobalStyle />
+      <ThemeProvider theme={createTheme(theme)}>
+        <CssBaseline />
+        <RouterProvider router={Router}>
           <NavBar></NavBar>
-          <Router />
-        </ThemeProvider>
-      </BrowserRouter>
+        </RouterProvider>
+      </ThemeProvider>
     </StylesProvider>
   )
 }
