@@ -1,7 +1,7 @@
 const isDST = d => {
   let jan = new Date(d.getFullYear(), 0, 1).getTimezoneOffset()
   let jul = new Date(d.getFullYear(), 6, 1).getTimezoneOffset()
-  return Math.max(jan, jul) != d.getTimezoneOffset()
+  return Math.max(jan, jul) !== d.getTimezoneOffset()
 }
 
 export default isDST
